+++ Okta/okta_modules/account_validator.py
@@ -1,10 +1,14 @@
-from sekoia_automation.account_validator import AccountValidator
-from okta.client import Client as OktaClient
-from functools import cached_property
 import asyncio
+from functools import cached_property
+
+from okta.client import Client as OktaClient
+from sekoia_automation.account_validator import AccountValidator
+
+from okta_modules import OktaModule
 
 
 class OktaAccountValidator(AccountValidator):
+    module: OktaModule
 
     @cached_property
     def client(self) -> OktaClient:
@@ -17,13 +21,15 @@ class OktaAccountValidator(AccountValidator):
             "orgUrl": self.module.configuration.base_url,
             "token": self.module.configuration.apikey,
         }
+
         return OktaClient(config)
 
-    def validate(self):
+    def validate(self) -> bool:
         try:
             loop = asyncio.get_event_loop()
             loop.run_until_complete(self.client.list_users())
         except Exception as e:
             self.log(f"Error while validating account: {e}", level="error")
             return False
+
         return True