+++ Okta/okta_modules/system_log_trigger.py
@@ -3,11 +3,13 @@ import time
 from collections.abc import Generator
 from datetime import datetime, timedelta, timezone
 from functools import cached_property
-from threading import Event
 from posixpath import join as urljoin
+from threading import Event
+from typing import Any, Optional
 
 import orjson
 import requests
+from cachetools import Cache, LRUCache
 from dateutil.parser import isoparse
 from sekoia_automation.connector import Connector, DefaultConnectorConfiguration
 from sekoia_automation.storage import PersistentJSON
@@ -40,24 +42,51 @@ class SystemLogConnector(Connector):
     module: OktaModule
     configuration: SystemLogConnectorConfiguration
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args: Any, **kwargs: Optional[Any]) -> None:
         super().__init__(*args, **kwargs)
         self._stop_event = Event()
         self.context = PersistentJSON("context.json", self._data_path)
         self.from_date = self.most_recent_date_seen
         self.fetch_events_limit = 1000
 
+        # Put cache size to 2000 in order to
+        self.cache_size = 2000
+        self.events_cache: Cache[str, bool] = self.load_events_cache()
+
         # Register signal to terminate thread
         signal.signal(signal.SIGINT, self.exit)
         signal.signal(signal.SIGTERM, self.exit)
 
-    def exit(self, _, __):
+    def load_events_cache(self) -> Cache[str, bool]:
+        """
+        Load the events cache.
+        """
+        cache: Cache[str, bool] = LRUCache(maxsize=self.cache_size)
+
+        with self.context as context:
+            # load the cache from the context
+            events_cache = context.get("events_cache", [])
+
+        for uuid in events_cache:
+            cache[uuid] = True
+
+        return cache
+
+    def save_events_cache(self) -> None:
+        """
+        Save the events cache.
+        """
+        with self.context as context:
+            # save the events cache to the context
+            context["events_cache"] = list(self.events_cache.keys())
+
+    def exit(self, _: Any, __: Optional[Any]) -> None:
         self.log(message="Stopping OKTA system logs connector", level="info")
         # Exit signal received, asking the processor to stop
         self._stop_event.set()
 
     @property
-    def most_recent_date_seen(self):
+    def most_recent_date_seen(self) -> datetime:
         now = datetime.now(timezone.utc)
 
         with self.context as cache:
@@ -78,12 +107,13 @@ class SystemLogConnector(Connector):
             return most_recent_date_seen
 
     @cached_property
-    def client(self):
+    def client(self) -> ApiClient:
         return ApiClient(
-            self.module.configuration.apikey, ratelimit_per_minute=self.configuration.ratelimit_per_minute
+            apikey=self.module.configuration.apikey,
+            ratelimit_per_minute=self.configuration.ratelimit_per_minute,
         )
 
-    def _handle_response_error(self, response: requests.Response):
+    def _handle_response_error(self, response: requests.Response) -> None:
         if not response.ok:
             message = (
                 f"Request on Okta API to fetch events failed with status {response.status_code} - {response.reason}"
@@ -98,9 +128,13 @@ class SystemLogConnector(Connector):
 
             raise FetchEventsException(message)
 
-    def __fetch_next_events(self, from_date: datetime) -> Generator[list, None, None]:
+    def __fetch_next_events(self, from_date: datetime) -> Generator[list[dict[str, Any]], None, None]:
         # set parameters
-        params = {"since": from_date.isoformat(), "limit": self.fetch_events_limit, "sortOrder": "ASCENDING"}
+        params: dict[str, str | int] = {
+            "since": from_date.isoformat(),
+            "limit": self.fetch_events_limit,
+            "sortOrder": "ASCENDING",
+        }
 
         # add optional parameters
         for param_name in ("filter", "q"):
@@ -121,10 +155,14 @@ class SystemLogConnector(Connector):
             # get events from the response
             events = response.json()
 
+            filtered_events = [
+                event for event in events if event.get("uuid") is not None and event["uuid"] not in self.events_cache
+            ]
+
             # yielding events if defined
-            if events:
-                INCOMING_MESSAGES.labels(intake_key=self.configuration.intake_key).inc(len(events))
-                yield events
+            if filtered_events:
+                INCOMING_MESSAGES.labels(intake_key=self.configuration.intake_key).inc(len(filtered_events))
+                yield filtered_events
             else:
                 logger.info(
                     f"The last page of events was empty. Waiting {self.configuration.frequency}s "
@@ -138,21 +176,17 @@ class SystemLogConnector(Connector):
 
             response = self.client.get(url, headers=headers)
 
-    def fetch_events(self) -> Generator[list, None, None]:
+    def fetch_events(self) -> Generator[list[dict[str, Any]], None, None]:
         most_recent_date_seen = self.from_date
 
         try:
             for next_events in self.__fetch_next_events(most_recent_date_seen):
                 if next_events:
                     # get the greater date seen in this list of events
-                    events_date = list(
-                        sorted(
-                            filter(
-                                lambda x: x is not None,
-                                map(lambda x: x["published"], next_events),
-                            )
-                        )
+                    events_date: list[str] = sorted(
+                        x["published"] for x in next_events if x.get("published") is not None
                     )
+
                     last_event_date = isoparse(events_date[-1])
 
                     # save the greater date ever seen
@@ -176,7 +210,7 @@ class SystemLogConnector(Connector):
         current_lag = now - most_recent_date_seen
         EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(int(current_lag.total_seconds()))
 
-    def next_batch(self):
+    def next_batch(self) -> None:
         # save the starting time
         batch_start_time = time.time()
 
@@ -192,6 +226,12 @@ class SystemLogConnector(Connector):
                 )
                 OUTCOMING_EVENTS.labels(intake_key=self.configuration.intake_key).inc(len(batch_of_events))
                 self.push_events_to_intakes(events=batch_of_events)
+
+                # Persist cache of event UUIDs after pushing to intake
+                for event in events:
+                    self.events_cache[event["uuid"]] = True
+
+                self.save_events_cache()
             else:
                 self.log(
                     message="No events to forward",
@@ -210,7 +250,7 @@ class SystemLogConnector(Connector):
             logger.debug(f"Next batch in the future. Waiting {delta_sleep} seconds")
             time.sleep(delta_sleep)
 
-    def run(self):
+    def run(self) -> None:
         self.log(message="Start fetching OKTA system logs", level="info")
 
         while not self._stop_event.is_set():