+++ Utils/CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 2025-10-02 - 1.4.5
+## 2025-12-02 - 1.4.6
 
 ### Changed
 
 - Add field for named timezones to Get Current Time action
 - Upgrade sekoia-automation-sdk
+
+## 2024-11-14 - 1.4.5
+
+### Added
+
+- Add the ability to define the accuracy of the time helper
+- Add logs to the Wait action
 
 ## 2024-11-04 - 1.4.4